Make plotly optional
